Break out firebase connection
