+++ starwarswiki/src/firebaseModel.js
@@ -1,0 +1,28 @@
+import { initializeApp, getDatabase, ref, get, set, onValue } from 'firebase/app';
+import { getAuth, GoogleAuthProvider, onAuthStateChanged, signInWithPopup } from 'firebase/auth';
+
+const app = initializeApp({
+	apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
+	authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
+	databaseURL: import.meta.env.VITE_FIREBASE_DATABASE_URL,
+	projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
+	storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
+	messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
+	appId: import.meta.env.VITE_FIREBASE_APP_ID,
+});
+
+const db = getDatabase(app);
+
+export default function readFirebase(path) {
+	return get(ref(db, path));
+}
+
+const provider = new GoogleAuthProvider();
+provider.setCustomParameters({
+	prompt: 'select_account ',
+});
+export const auth = getAuth();
+
+export function signInWithGooglePopup() {
+	return signInWithPopup(auth, provider);
+}
