+++ starwarswiki/src/main.jsx
@@ -1,32 +1,12 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import {QueryClient, QueryClientProvider} from 'react-query';
+import { QueryClient, QueryClientProvider } from 'react-query';
 import model from './model.js';
-import {configure, observable} from 'mobx';
-import {initializeApp} from 'firebase/app';
-import {getAuth, GoogleAuthProvider, onAuthStateChanged, signInWithPopup} from "firebase/auth";
+import { configure, observable } from 'mobx';
+import { connectFirebase } from './firebaseModel';
 
-const app = initializeApp({
-	apiKey: import.meta.env.VITE_FIREBASE_API_KEY,
-	authDomain: import.meta.env.VITE_FIREBASE_AUTH_DOMAIN,
-	databaseURL: import.meta.env.VITE_FIREBASE_DATABASE_URL,
-	projectId: import.meta.env.VITE_FIREBASE_PROJECT_ID,
-	storageBucket: import.meta.env.VITE_FIREBASE_STORAGE_BUCKET,
-	messagingSenderId: import.meta.env.VITE_FIREBASE_MESSAGING_SENDER_ID,
-	appId: import.meta.env.VITE_FIREBASE_APP_ID,
-});
-const provider = new GoogleAuthProvider();
-provider.setCustomParameters({
-	prompt: "select_account "
-});
-export const auth = getAuth();
-
-export function signInWithGooglePopup() {
-	return signInWithPopup(auth, provider);
-}
-
-configure({enforceActions: 'never'});
+configure({ enforceActions: 'never' });
 const reactiveModel = observable(model);
 
 onAuthStateChanged(auth, (user) => {
@@ -53,7 +33,7 @@ export const queryClient = new QueryClient({
 
 ReactDOM.createRoot(document.getElementById('root')).render(
 	<QueryClientProvider client={queryClient}>
-		<App model={reactiveModel}/>
+		<App model={reactiveModel} />
 	</QueryClientProvider>,
 );
